% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{find_unique_patterns}
\alias{find_unique_patterns}
\title{Find unique patterns in matrix}
\usage{
find_unique_patterns(data)
}
\arguments{
\item{data}{A list of gene vectors}
}
\value{
A matrix with counts per pattern
}
\description{
Find unique patterns in matrix
}
\examples{
\dontrun{
mdp <- find_unique_patterns(data)
}
}
