% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{missing_pattern_pval_kw}
\alias{missing_pattern_pval_kw}
\title{Kruskal-Wallis test to generate a per cell p-value based on missing data patterns}
\usage{
missing_pattern_pval_kw(x, mdp)
}
\arguments{
\item{x}{Item from list of cell gene expression vectors}

\item{mdp}{Matrix with counts per pattern}
}
\value{
list with a p-value for each cell
}
\description{
Kruskal-Wallis test to generate a per cell p-value based on missing data patterns
}
\examples{
\dontrun{
pvalue.list.KW <- unlist(lapply(cl, FUN = missing_pattern_pval_kw, mdp = mdp))
}

}
