% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAR-package.R
\docType{package}
\name{OAR-package}
\alias{OAR}
\alias{OAR-package}
\title{OAR: Scoring transcriptional shifts in scRNAseq}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

OAR (observed at random) score is a measure of transcriptional shifts among cells, allowing cell prioritization for downstream applications. For best results, test a group of similar cells where you expect some heterogeneity - i.e. one cell type across various biological samples or conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sanin-Lab/OARscRNA}
  \item \url{https://Sanin-Lab.github.io/OARscRNA/}
  \item \url{https://sanin-lab.github.io/OARscRNA/}
  \item Report bugs at \url{https://github.com/Sanin-Lab/OARscRNA/issues}
}

}
\author{
\strong{Maintainer}: David Sanin \email{david.sanin@jhu.edu} (\href{https://orcid.org/0000-0003-0188-7267}{ORCID})

Authors:
\itemize{
  \item Ruizhe Chen \email{rzchen@jhu.edu} (\href{https://orcid.org/0000-0003-3924-3328}{ORCID})
  \item Hannah Moore \email{hmoore14@jhmi.edu} (\href{https://orcid.org/0009-0006-8245-5368}{ORCID})
  \item Dmitrijs Lvovs \email{dlvovs1@jh.edu} (\href{https://orcid.org/0009-0003-2152-6853}{ORCID})
  \item Paul Gueguen \email{paul.gueguen@fgcz.ethz.ch} (\href{https://orcid.org/0000-0003-2930-6073}{ORCID})
}

}
\keyword{internal}
