% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oar_functions.R
\name{oar_by_factor}
\alias{oar_by_factor}
\title{Generate OAR score within each cluster and add them to full objects metadata}
\usage{
oar_by_factor(
  data,
  count.filter = 1,
  blacklisted.genes = NULL,
  suffix = "",
  tolerance = TRUE,
  factor = "ident",
  cores = 1
)
}
\arguments{
\item{data}{a Seurat (v5) object.}

\item{count.filter}{a numeric value indicating the minimum fraction of cells expressing any given gene that will be included in the analysis, default is 1. Values between 0.5 and 2 are recommended.}

\item{blacklisted.genes}{a character vector with gene names to be excluded from the analysis. Default is empty.}

\item{suffix}{a string to append to the output variables. Default is empty}

\item{factor}{a  character string mapping to a column in \code{meta.data}. Defaults to \code{ident}.}

\item{cores}{a numeric value indicating the number of cores to use un parallel processing. Use \code{parallel::detectCores()} or \code{parallelly::availableCores()} to identify possibilities.}
}
\value{
a Seurat object with OAR stats added into meta data
}
\description{
Generate OAR score within each cluster and add them to full objects metadata
}
\examples{
\dontrun{
pmbcs <- oar_by_cluster(pmbcs)
}
}
