% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{oar_exact_missing_data_patterns}
\alias{oar_exact_missing_data_patterns}
\title{Identify missing data patterns with exact matches and combine all unique missing data patterns}
\usage{
oar_exact_missing_data_patterns(data)
}
\arguments{
\item{data}{a minimal dataset ready for processing.}
}
\value{
missing data pattern vector
}
\description{
Identify missing data patterns with exact matches and combine all unique missing data patterns
}
\examples{
\dontrun{
mdp <- oar_exact_missing_data_patterns(data)
}
}
