% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oar_functions.R
\name{oar_fold}
\alias{oar_fold}
\title{Generate OAR fold stats for Seurat object and add them to meta data, run iteratively with different partitions of data and average across them.}
\usage{
oar_fold(
  data,
  seurat_v5 = T,
  count.filter = 1,
  blacklisted.genes = NULL,
  gene.ratio = 20,
  iterations = 10,
  parallel.loop = T,
  cores = 12,
  suffix = ""
)
}
\arguments{
\item{data}{A Seurat (v5) object or a matrix with cell barcodes as column names and genes as row names.}

\item{seurat_v5}{A Boolean to indicate if supplied data is a Seurat object, default is TRUE}

\item{count.filter}{A numeric value indicating the minimum fraction of cells expressing any given gene that will be included in the analysis, default is 0.01. Values between 0.005 and 0.02 are recommended.}

\item{blacklisted.genes}{A character vector with gene names to be excluded from the analysis. Default is empty.}

\item{gene.ratio}{A numeric value indicating the ratio of genes to cells to partition data matrix. Default is 20. Values between 15 and 25 are recommended.}

\item{iterations}{A numeric value indicating the number of times to iterate the OARscore calculation. Default and recommendation is 10. Can use 1 to not iterate the calculation.}

\item{parallel.loop}{A Boolean to indicate if process should be run in parallel cores. Currently only \code{doParallel} and \code{foreach} are supported, default is FALSE}

\item{cores}{A numeric value indicating the number of cores to use un parallel processing. Use \code{detectCores()} to identify possibilities. Default is 12. Ignored if \code{parallel.loop} set to FALSE.}

\item{suffix}{A string to append to the output variables. Default is empty}
}
\value{
A Seurat object with OAR stats added into meta data
}
\description{
Generate OAR fold stats for Seurat object and add them to meta data, run iteratively with different partitions of data and average across them.
}
\examples{
\dontrun{
pmbcs <- oar_fold(pmbcs)}
}
