% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{oar_hamming_distance}
\alias{oar_hamming_distance}
\title{Calculate hamming distances between genes}
\usage{
oar_hamming_distance(data, cores = 1)
}
\arguments{
\item{data}{a minimal dataset ready for processing.}

\item{cores}{a numeric value indicating the number of cores to use un parallel processing. Use \code{parallel::detectCores()} or \code{parallelly::availableCores()} to identify possibilities.}
}
\value{
A list of matrices of hamming distances across gene bins
}
\description{
Calculate hamming distances between genes
}
\examples{
\dontrun{
##Automatic tolerance setting
dm <- oar_hamming_distance(data, cores = 2)

}
}
