% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oar_missing_data_graph}
\alias{oar_missing_data_graph}
\title{Group missing data patterns based on tolerance with a graph}
\usage{
oar_missing_data_graph(dm, tol = 0.05)
}
\arguments{
\item{dm}{a matrix of gene vector hamming distances.}

\item{tol}{a numeric value indicating the maximum fraction of mismatch in genes to group as a pattern, default is 0.05.}
}
\value{
missing data pattern vector
}
\description{
Group missing data patterns based on tolerance with a graph
}
\examples{
\dontrun{
mdp <- oar_missing_data_patterns(dm, tol = 0.05)
}
}
