% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{oar_missing_data_patterns}
\alias{oar_missing_data_patterns}
\title{Identify missing data patterns allowing for mismatch}
\usage{
oar_missing_data_patterns(data, dm)
}
\arguments{
\item{data}{a minimal dataset ready for processing.}

\item{dm}{a matrix of hamming distances between gene vectors.}
}
\value{
Vector of missing data patterns
}
\description{
Identify missing data patterns allowing for mismatch
}
\examples{
\dontrun{
mdp <- oar_missing_data_patterns(data, dm)
}
}
