% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{oar_patterns}
\alias{oar_patterns}
\title{Identify gene co-expression patterns allowing for mismatch}
\usage{
oar_patterns(data, dm)
}
\arguments{
\item{data}{a minimal dataset ready for processing.}

\item{dm}{a matrix of hamming distances between gene vectors.}
}
\value{
Vector of gene co-expression patterns
}
\description{
Identify gene co-expression patterns allowing for mismatch
}
\examples{
\dontrun{
mdp <- oar_patterns(data, dm)
}
}
