% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oar_preprocess_data}
\alias{oar_preprocess_data}
\title{Prepare data for oar fold functions}
\usage{
oar_preprocess_data(data, tr = 1, seurat_v5 = TRUE, blacklisted.genes = NULL)
}
\arguments{
\item{data}{a seurat object or gene expression matrix}

\item{tr}{a filtering threshold. Default is 1.}

\item{seurat_v5}{a boolean to indicate if supplied data is a Seurat object, default is TRUE}

\item{blacklisted.genes}{a character vector with gene names to be excluded from the analysis. Default is empty.}
}
\value{
Data matrix with blacklisted genes removed
}
\description{
Prepare data for oar fold functions
}
\examples{
\dontrun{
data <- oar_preprocess_data(data)
}
}
