% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_gene_expression.R
\name{oardeg}
\alias{oardeg}
\title{Generate DEGs based on OAR score}
\usage{
oardeg(
  data,
  seurat_v5 = T,
  score = NULL,
  count.filter = 1,
  splines = TRUE,
  degrees.freedom = 5,
  method = "glmGamPoi",
  blacklisted.genes = NULL,
  ncores = 12,
  auto.threshold = TRUE,
  custom.tr = NULL
)
}
\arguments{
\item{data}{A Seurat (v5) object or a data.frame with cell barcodes as column names and genes as row names. Seurat object must have a column in metadata with OARscore.}

\item{seurat_v5}{A Boolean to indicate if supplied data is a Seurat object, default is TRUE}

\item{score}{A named numeric vector with OARscores to be used in the model. Names should be cell barcodes. Ignored if Seurat object is supplied.}

\item{count.filter}{A numeric value indicating the minimum fraction of cells expressing any given gene that will be included in the analysis, default is 0.01. Values between 0.005 and 0.02 are recommended.}

\item{splines}{A Boolean to indicate if splines should be used to fit model, default is TRUE}

\item{degrees.freedom}{A numeric value indicating the degrees of freedom to calculate splines. Default is 5. Values between 3 and 5 are recommended.}

\item{method}{A character with one of \verb{"glmGamPoi","DESeq2","edgeR"} to indicate preferred method. glmGamPoi is default.}

\item{blacklisted.genes}{A character vector with gene names to be excluded from the analysis. Default is empty.}

\item{ncores}{A numeric value indicating the number of cores to use un parallel processing. Use \code{detectCores()} to identify possibilities. Default is 2. Ignored if \code{parallel.loop} set to FALSE.}

\item{auto.threshold}{A Boolean to indicate if FDR threshold should be calculated from the data, default is TRUE}

\item{custom.tr}{A numeric value to use as an FDR threshold. Ignored if \code{auto.threshold} is set to TRUE.}
}
\value{
A dataframe with p-values for genes that significantly contribute to OAR score.
}
\description{
Generate DEGs based on OAR score
}
\examples{
\dontrun{
degs <- oardeg(data)
}
}
