% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oaridentifyfolds}
\alias{oaridentifyfolds}
\title{Identify number of folds to use for data based on desired gene-cell ratio}
\usage{
oaridentifyfolds(data, gene.ratio, max.cells)
}
\arguments{
\item{data}{A matrix with low expression genes filtered}

\item{gene.ratio}{A numeric value indicating the ratio of genes to cells to partition data matrix. Default is 20. Values between 15 and 25 are recommended.}

\item{max.cells}{Total number of cells available in data}
}
\value{
A variable called folds that contains a numeric value
}
\description{
Identify number of folds to use for data based on desired gene-cell ratio
}
\examples{
\dontrun {
folds <- oaridentifyfolds(data)
}
}
