% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oarpreprocessdata}
\alias{oarpreprocessdata}
\title{Prepare data for oar fold functions}
\usage{
oarpreprocessdata(data, tr, seurat_v5 = T, blacklisted.genes = NULL)
}
\arguments{
\item{data}{Seurat object or gene expression matrix}

\item{tr}{Filtering threshold}

\item{seurat_v5}{A Boolean to indicate if supplied data is a Seurat object, default is TRUE}

\item{blacklisted.genes}{A character vector with gene names to be excluded from the analysis. Default is empty.}
}
\value{
Data matrix with blacklisted gens removed
}
\description{
Prepare data for oar fold functions
}
\examples{
\dontrun{
data <- oarpreprocessdata(data)
}
}
