% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pattern_pval_kw}
\alias{pattern_pval_kw}
\title{Kruskal-Wallis test to generate a per cell p-value based on gene co-expression patterns}
\usage{
pattern_pval_kw(x, cgp)
}
\arguments{
\item{x}{Item from list of cell gene expression vectors}

\item{cgp}{Matrix with gene participation per pattern}
}
\value{
list with a p-value for each cell
}
\description{
Kruskal-Wallis test to generate a per cell p-value based on gene co-expression patterns
}
\examples{
\dontrun{
pvalue <- pattern_pval_kw(x, cgp)
}

}
