% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatter_score_missing}
\alias{scatter_score_missing}
\title{Create scatter plot of OAR score vs percent missing}
\usage{
scatter_score_missing(
  data,
  group.by = "seurat_clusters",
  seurat_v5 = TRUE,
  suffix = ""
)
}
\arguments{
\item{data}{A seurat v5 object that has oar score in meta data.}

\item{group.by}{meta data category to color data by. Default is seurat_clusters.}

\item{seurat_v5}{A Boolean to indicate if supplied data is a Seurat object, default is TRUE}

\item{suffix}{A string that was previously appended to the output variables. Default is empty.}
}
\value{
Scatter plot of OAR score vs percent missing, colored by grouping of choice.
}
\description{
Create scatter plot of OAR score vs percent missing
}
\examples{
\dontrun{
pmbcs_oar <- scatter_score_missing(pmbcs_oar)
}
}
