% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatter_score_sd}
\alias{scatter_score_sd}
\title{Create scatter plot of OAR score vs OAR score Standard Deviation}
\usage{
scatter_score_sd(data, seurat_v5 = TRUE)
}
\arguments{
\item{data}{A seurat v5 object that has oar score in meta data.}

\item{seurat_v5}{A Boolean to indicate if supplied data is a Seurat object, default is TRUE}
}
\value{
Scatter plot of OAR score standard deviation vs OAR score, colored by percent missingness of each cell.
}
\description{
Create scatter plot of OAR score vs OAR score Standard Deviation
}
\examples{
\dontrun{
pmbcs_oar <- scatter_score_sd(pmbcs_oar)
}
}
