% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{get_missing_pattern_genes}
\alias{get_missing_pattern_genes}
\title{Identify missing data patterns allowing for mismatch}
\usage{
get_missing_pattern_genes(data)
}
\arguments{
\item{data}{a Seurat object that has had oar() run on it previously.}
}
\value{
Dataframe of variable genes and the missing data pattern they participate in.
}
\description{
Identify missing data patterns allowing for mismatch
}
\examples{
\dontrun{
mdp <- get_missing_pattern_genes(data)
}
}
