% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_patterns.R
\name{get_pattern_genes}
\alias{get_pattern_genes}
\title{Create list of which genes participate in each pattern.}
\usage{
get_pattern_genes(data)
}
\arguments{
\item{data}{a Seurat object that has had \code{oar()} or \code{oar_by_cluster()} run on it previously.}
}
\value{
data.frame of genes annotated with gene co-expression pattern they participate in (globally or by cluster).
}
\description{
Create list of which genes participate in each pattern.
}
\examples{
\dontrun{
mdp <- get_pattern_genes(data)
}
}
