% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAR-package.R
\docType{package}
\name{OAR-package}
\alias{OAR}
\alias{OAR-package}
\title{OAR: Heterogeneity Scoring on scRNA Sequencing Data Based on Missingness}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Sanin \email{david.sanin@jhu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
