% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oar_functions.R
\name{oar_base}
\alias{oar_base}
\title{Generate scores and p-values to determine transcriptional shifts}
\usage{
oar_base(data, cgp)
}
\arguments{
\item{data}{A gene-cell expression matrix with NA values in place of 0s.}

\item{cgp}{A vector indicating the pattern to which each gene belongs.}
}
\value{
Data frame with OAR-score, p-value, adjusted p-value, and sparsity as a percentage for each cell.
}
\description{
Generate scores and p-values to determine transcriptional shifts
}
\examples{
\dontrun{
output <- oar_base(data, cgp)
}
}
