% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oar_functions.R
\name{oar_by_cluster}
\alias{oar_by_cluster}
\title{Generate OAR score within each cluster and add them to full objects metadata}
\usage{
oar_by_cluster(
  data,
  count.filter = 1,
  blacklisted.genes = NULL,
  suffix = "",
  tolerance = TRUE,
  cores = 1
)
}
\arguments{
\item{data}{a clustered Seurat (v5) object.}

\item{count.filter}{a numeric value indicating the minimum fraction of cells expressing any given gene that will be included in the analysis, default is 1. Values between 0.5 and 2 are recommended.}

\item{blacklisted.genes}{a character vector with gene names to be excluded from the analysis. Default is empty.}

\item{suffix}{a string to append to the output variables. Default is empty}

\item{tolerance}{a boolean or numeric value controlling the tolerance threshold for pattern matching. If set to \code{TRUE}, then tolerance is automatically adjusted. Alternatively, user may supply a numeric value indicating the maximum fraction of mismatch between pairs of genes for pattern grouping. Values between 0.01 and 0.05 are recommended.}

\item{cores}{a numeric value indicating the number of cores to use un parallel processing. Use \code{parallel::detectCores()} or \code{parallelly::availableCores()} to identify possibilities.}
}
\value{
a Seurat object with OAR stats added into meta data
}
\description{
Generate OAR score within each cluster and add them to full objects metadata
}
\examples{
\dontrun{
pmbcs <- oar_by_cluster(pmbcs)
}
}
