% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oar_gene_graph}
\alias{oar_gene_graph}
\title{Group gene co-expression patterns based on tolerance with a graph}
\usage{
oar_gene_graph(dm, tol)
}
\arguments{
\item{dm}{a matrix of gene vector hamming distances.}

\item{tol}{a numeric value indicating the maximum fraction of mismatch in genes to group as a pattern.}
}
\value{
gene co-expression pattern vector
}
\description{
Group gene co-expression patterns based on tolerance with a graph
}
\examples{
\dontrun{
cgp <- oar_gene_graph(dm, tol)
}
}
