% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{oar_missing_data_plot}
\alias{oar_missing_data_plot}
\title{Plot identified missing data patterns}
\usage{
oar_missing_data_plot(data, mdp = NULL, seurat_v5 = TRUE)
}
\arguments{
\item{data}{a gene-cell expression matrix with NA values in place of 0s and 1s everywhere else or a Seurat Object to which \code{oar} has been applied to.}

\item{mdp}{a vector indicating the pattern to which each gene belongs. Default is NULL.}

\item{seurat_v5}{a boolean to indicate where or not the input data is a Seurat object. Default is TRUE.}
}
\value{
Plot of missing data patterns.
}
\description{
Plot identified missing data patterns
}
\examples{
\dontrun{
##Starting from a Seurat Object analysed jointly
output <- oar_missing_data_plot(data, seurat_v5 = TRUE)

##Starting from filtered and binarized expression matrix
output <- oar_missing_data_plot(data, mdp, seurat_v5 = FALSE)
}
}
